refactor: extract useMarkupEditor hook from App.tsx (W0-T008)
Moves the Tiptap editor construction, extension registration, save
debounce, and user-edit detection loop into a new
`src/hooks/use-markup-editor.ts`. `App.tsx` now calls
`useMarkupEditor(options)` and receives `{ editor, editorRef,
lastDocSnapshot }`.

Behavior is identical: same extensions (`StarterKit`, `Placeholder`,
`AgentCursors`, `DocMinimap`), same placeholder text, same save and
user-edit debounce timings (read from the existing module-scope
constants in App.tsx and threaded in as options so nobody has to chase
two copies).

Session state, message forwarding, and task callbacks remain in App.tsx
— those belong to W0-T009 and W0-T010. The hook reads session
getter/setter callbacks through a small `callbacksRef` bag so App.tsx
can create the editor before `useSession` has produced its setters
(`useSession` itself still needs the editor for document hydration).

- src/App.tsx: -93/+30 lines
- src/hooks/use-markup-editor.ts: new (173 lines)
